+++ frontend/src/router/router.ts
@@ -19,7 +19,8 @@ export enum RouteName {
   Profile = 'Profile',
   // Friends = 'Friends',
 
-  Beta = 'Beta',
+  Credits = 'Credits',
+  About = 'About',
 
   Admin = 'Admin',
 
@@ -40,9 +41,8 @@ const DishInfographicPage = () => import('@/pages/dish-infographic/dish-infograp
 const SettingsPage = () => import('@/pages/settings/settings-page.vue')
 const ProfilePage = () => import('@/pages/profile-page.vue')
 // const FriendsPage = () => import('@/pages/friends/friends-page.vue')
-
-// Misc
-const BetaPage = () => import('@/pages/beta/beta.vue')
+const CreditsPage = () => import('@/pages/credits/credits-page.vue')
+const AboutPage = () => import('@/pages/about/about-page.vue')
 
 // Admin
 const AdminPage = () => import('@/pages/admin/admin.vue')
@@ -141,9 +141,14 @@ const router = createRouter({
     //   component: FriendsPage
     // },
     {
-      path: '/beta',
-      name: RouteName.Beta,
-      component: BetaPage
+      path: '/credits',
+      name: RouteName.Credits,
+      component: CreditsPage
+    },
+    {
+      path: '/about',
+      name: RouteName.About,
+      component: AboutPage
     },
     {
       path: '/admin',